# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Plataforma de Auditoria Fiscal Preventiva (CFDI Intelligence)
  version: 1.0.0
  description: |
    API REST para la plataforma de deteccion automatica de riesgos fiscales.
    Permite gestionar empresas, ingestar CFDIs y estados de cuenta bancarios,
    conciliar movimientos, detectar riesgos fiscales, calcular scoring de
    cumplimiento y descargar CFDIs directamente del SAT via FIEL (e.firma).

    **Autenticacion**: Todos los endpoints (excepto /auth/register, /auth/login
    y /health) requieren un Bearer Token JWT obtenido al autenticarse.

    **Vencimiento del token**: 8 horas desde su emision.
servers:
  - url: https://api.cfdi-intelligence.com
    description: Produccion
  - url: http://localhost:8000
    description: Desarrollo local
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/empresas/{empresa_id}/riesgos:
    get:
      operationId: listarRiesgos
      tags: [Riesgos]
      summary: Listar riesgos fiscales detectados
      description: |
        Retorna las detecciones de riesgo fiscal para la empresa, ordenadas
        por severidad (critico > alto > medio > bajo) y fecha de deteccion.

        Se puede filtrar por periodo, severidad y estado.
      parameters:
        - $ref: '#/components/parameters/EmpresaId'
        - name: periodo
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          description: Filtrar por periodo YYYY-MM
          example: "2024-01"
        - name: severidad
          in: query
          schema:
            type: string
            enum: [critico, alto, medio, bajo]
          description: Filtrar por nivel de severidad
        - name: estado
          in: query
          schema:
            type: string
            enum: [abierto, en_revision, en_espera_cfdi, confirmado, resuelto, descartado, pendiente]
            default: abierto
          description: Filtrar por estado de la deteccion
      responses:
        '200':
          description: Lista de riesgos detectados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiesgosResponse'
              example:
                total: 3
                riesgos:
                  - id: "d1e2f3a4-0001-0001-0001-000000000001"
                    codigo: "R001"
                    nombre: "Ingreso sin CFDI"
                    severidad: critico
                    monto_afectado: 50000.00
                    descripcion: "Deposito bancario sin comprobante fiscal correspondiente"
                    estado: abierto
                    periodo: "2024-01"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/riesgos/{riesgo_id}/resolver:
    patch:
      operationId: resolverRiesgo
      tags: [Riesgos]
      summary: Marcar riesgo como resuelto
      description: |
        Marca una deteccion de riesgo como resuelta. Opcionalmente se pueden
        agregar notas de resolucion.
      parameters:
        - name: riesgo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID de la deteccion de riesgo
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                notas:
                  type: string
                  description: Notas sobre la resolucion del riesgo
            example:
              notas: "Se emitio el CFDI faltante con folio 2024-0123"
      responses:
        '200':
          description: Riesgo marcado como resuelto
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensaje:
                    type: string
                  riesgo_id:
                    type: string
                  estado:
                    type: string
              example:
                mensaje: "Riesgo marcado como resuelto"
                riesgo_id: "d1e2f3a4-0001-0001-0001-000000000001"
                estado: resuelto
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ============================================================
  # PATRONES DE RIESGO — Búsqueda híbrida (AgentDB-style)
  # ============================================================
  /api/v1/riesgos/patrones/buscar:
    get:
      operationId: buscarPatronesRiesgo
      tags: [Riesgos]
      summary: Búsqueda híbrida de patrones de riesgo
      description: |
        Recupera patrones de riesgo similares usando búsqueda híbrida:
        70% similitud textual (pg_trgm) + 30% confianza del patrón + filtros de metadata.

        Opcionalmente aplica MMR (Maximal Marginal Relevance) para maximizar
        la diversidad de resultados y evitar redundancia.

        Incluye síntesis de contexto: tasa de éxito histórica, resoluciones
        frecuentes y severidad predominante.
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Texto de búsqueda (descripción del riesgo a buscar)
          example: "depósito bancario sin CFDI de ingreso correspondiente"
        - name: tipo_riesgo
          in: query
          required: false
          schema:
            type: string
            enum: [INGRESO_NO_FACTURADO, GASTO_SIN_CFDI, CFDI_NO_COBRADO, CFDI_NO_PAGADO,
                   DIFERENCIA_IVA, RFC_INVALIDO, CFDI_CANCELADO_COBRADO, DIFERENCIA_TIPO_CAMBIO]
          description: Filtrar por tipo de riesgo
        - name: severidad
          in: query
          required: false
          schema:
            type: string
            enum: [critico, alto, medio, bajo]
          description: Filtrar por severidad
        - name: categoria
          in: query
          required: false
          schema:
            type: string
          description: Filtrar por categoría fiscal (IVA, ISR, COMPROBACION, etc.)
        - $ref: '#/components/parameters/EmpresaId'
        - name: mmr
          in: query
          required: false
          schema:
            type: boolean
            default: true
          description: Aplicar MMR para diversificar resultados
        - name: lambda_mmr
          in: query
          required: false
          schema:
            type: number
            minimum: 0
            maximum: 1
            default: 0.5
          description: "Balance MMR: 0=máxima relevancia, 1=máxima diversidad"
        - name: k
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
          description: Número máximo de patrones a retornar
      responses:
        '200':
          description: Patrones similares con contexto sintetizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  patrones:
                    type: array
                    items:
                      $ref: '#/components/schemas/RiskPattern'
                  contexto:
                    $ref: '#/components/schemas/ContextoSintetizado'
              example:
                total: 3
                patrones:
                  - id: "a1b2c3d4-0001-0001-0001-000000000001"
                    tipo_riesgo: "INGRESO_NO_FACTURADO"
                    severidad: "critico"
                    categoria: "ISR/IVA"
                    descripcion: "Depósito de $45,000 sin CFDI de ingreso"
                    resolucion: "Se emitió CFDI retroactivo con folio 2024-0123"
                    confianza: 0.95
                    uso_count: 12
                    similitud: 0.8734
                  - id: "a1b2c3d4-0002-0002-0002-000000000002"
                    tipo_riesgo: "INGRESO_NO_FACTURADO"
                    severidad: "critico"
                    categoria: "ISR/IVA"
                    descripcion: "Transferencia recibida $32,500 sin factura"
                    resolucion: "Ingreso correspondía a préstamo de socio, se documentó"
                    confianza: 0.88
                    uso_count: 5
                    similitud: 0.7621
                contexto:
                  total_patrones: 3
                  tasa_exito: 0.93
                  severidad_predominante: "critico"
                  resoluciones_frecuentes:
                    - "Se emitió CFDI retroactivo con folio 2024-0123"
                    - "Ingreso correspondía a préstamo de socio, se documentó"
                  resumen: "Basado en 3 casos similares: 93% resueltos exitosamente. Severidad predominante: critico. Resolución más frecuente: Se emitió CFDI retroactivo."
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/riesgos/{riesgo_id}/guardar-patron:
    post:
      operationId: guardarPatronRiesgo
      tags: [Riesgos]
      summary: Guardar detección resuelta como patrón
      description: |
        Persiste una detección resuelta como patrón reutilizable en la base de
        conocimiento. Alimenta la búsqueda híbrida para casos futuros similares.

        Se recomienda llamar este endpoint al resolver o descartar una detección
        para que el sistema aprenda de cada resolución.
      security:
        - BearerAuth: []
      parameters:
        - name: riesgo_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID de la detección a guardar como patrón
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolucion]
              properties:
                resolucion:
                  type: string
                  description: Descripción de cómo se resolvió el riesgo
                fue_falso_positivo:
                  type: boolean
                  default: false
                  description: Marcar si era un falso positivo para excluir de búsquedas
                confianza:
                  type: number
                  minimum: 0
                  maximum: 1
                  default: 1.0
                  description: Confianza en la calidad de esta resolución (0.0–1.0)
            example:
              resolucion: "Depósito correspondía a anticipo de cliente. Se emitió CFDI tipo ingreso folio A-2024-089."
              fue_falso_positivo: false
              confianza: 0.95
      responses:
        '200':
          description: Patrón guardado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  patron_id:
                    type: string
                    format: uuid
                  mensaje:
                    type: string
              example:
                patron_id: "f9e8d7c6-0001-0001-0001-000000000001"
                mensaje: "Patrón guardado correctamente"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ============================================================
  # ACCIONES
  # ============================================================
  /api/v1/acciones/{deteccion_id}/ejecutar:
    post:
      operationId: ejecutarAccion
      tags: [Acciones]
      summary: Ejecutar accion sobre una deteccion
      description: |
        Ejecuta una accion de flujo de trabajo sobre una deteccion de riesgo
        y actualiza su estado segun la accion:

        | Accion | Estado resultante |
        |--------|------------------|
        | marcar_revisado | en_revision |
        | solicitar_cfdi | en_espera_cfdi |
        | emitir_cfdi | en_espera_cfdi |
        | confirmar_match | confirmado |
        | descartar | descartado |
        | resolver | resuelto |
      parameters:
        - name: deteccion_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID de la deteccion sobre la que se ejecuta la accion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccionRequest'
            example:
              tipo: marcar_revisado
              notas: "Pendiente de verificacion con el cliente"
      responses:
        '200':
          description: Accion ejecutada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccionResponse'
              example:
                deteccion_id: "d1e2f3a4-0001-0001-0001-000000000001"
                estado_anterior: abierto
                estado_nuevo: en_revision
        '400':
          description: Tipo de accion desconocido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Tipo de accion desconocido: accion_invalida"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ============================================================
  # SCORING
  # ============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido al registrarse o hacer login.
        Incluir en el header: Authorization: Bearer {token}
        Duracion: 8 horas.
  parameters:
    EmpresaId:
      name: empresa_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: ID unico de la empresa (UUID)
      example: "a1b2c3d4-0001-0001-0001-000000000001"
  responses:
    UnauthorizedError:
      description: Token JWT ausente, invalido o expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Se requiere autenticacion"
    ForbiddenError:
      description: El usuario no tiene acceso a este recurso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Sin acceso a esta empresa"
    NotFoundError:
      description: El recurso solicitado no existe
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Empresa no encontrada"
...
  schemas:
    RiskPattern:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tipo_riesgo:
          type: string
          description: Código del tipo de riesgo
        severidad:
          type: string
          enum: [critico, alto, medio, bajo]
        categoria:
          type: string
          description: Categoría fiscal (IVA, ISR, COMPROBACION, etc.)
        descripcion:
          type: string
          description: Descripción del caso de riesgo
        resolucion:
          type: string
          nullable: true
          description: Cómo se resolvió este caso
        confianza:
          type: number
          minimum: 0
          maximum: 1
          description: Calidad del patrón (0.0–1.0)
        uso_count:
          type: integer
          description: Número de veces que se ha consultado este patrón
        similitud:
          type: number
          description: Score de similitud con la búsqueda (0.0–1.0)
    ContextoSintetizado:
      type: object
      properties:
        total_patrones:
          type: integer
        tasa_exito:
          type: number
          description: Porcentaje de casos resueltos exitosamente (0.0–1.0)
        severidad_predominante:
          type: string
          enum: [critico, alto, medio, bajo, ""]
        resoluciones_frecuentes:
          type: array
          items:
            type: string
          description: Top 3 resoluciones más frecuentes
        resumen:
          type: string
          description: Resumen narrativo generado automáticamente
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          description: Descripcion del error
...
    Riesgo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        codigo:
          type: string
          description: Codigo unico del tipo de riesgo (ej. R001)
        nombre:
          type: string
          description: Nombre descriptivo del riesgo
        severidad:
          type: string
          enum: [critico, alto, medio, bajo]
        monto_afectado:
          type: number
          format: float
          nullable: true
        descripcion:
          type: string
          nullable: true
        cfdi_id:
          type: string
          format: uuid
          nullable: true
        movimiento_id:
          type: string
          format: uuid
          nullable: true
        estado:
          type: string
          enum: [abierto, pendiente, en_revision, en_espera_cfdi, confirmado, resuelto, descartado]
        periodo:
          type: string
          example: "2024-01"
        created_at:
          type: string
          format: date-time
    RiesgosResponse:
      type: object
      properties:
        total:
          type: integer
        riesgos:
          type: array
          items:
            $ref: '#/components/schemas/Riesgo'
    AccionRequest:
      type: object
      required: [tipo]
      properties:
        tipo:
          type: string
          enum: [marcar_revisado, solicitar_cfdi, emitir_cfdi, confirmar_match, descartar, resolver]
          description: Tipo de accion a ejecutar
        notas:
          type: string
          nullable: true
          description: Notas opcionales sobre la accion
    AccionResponse:
      type: object
      properties:
        deteccion_id:
          type: string
          format: uuid
        estado_anterior:
          type: string
        estado_nuevo:
          type: string
...